pyproject.toml
[project]
name = "ai-kaeyris"
version = "0.0.1"
description = "Personal AI sandbox platform — modular FastAPI services on Raspberry Pi"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "argon2-cffi",
    "redis>=5.0",
    "rq>=2.0",
    "openai>=1.50",
    "sqlalchemy>=2.0",
    "alembic",
    "aiosqlite",
    "python-multipart",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
    "fakeredis",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
